ADD support for multiple OAuth tokens
If a user tries to create a new token while he already has one, token
creation will fail and gist will be posted as anonymously.
This is because the `note` attribute in posted json cannot already
exists in the list of personal access tokens.

In order to bypass it, create a new note with timestamp to ensure
uniqueness.

Related #34

Details
* ADD ability to have multiple tokens based on timestamp
* UPDATE documentation
* REFACTOR `OAuth` authentication
* UPDATE changelog